🐛 Load entrypoint plugins only when APIs are called
