+++ datar/all.py
@@ -2,6 +2,7 @@
 
 _locs = locals()
 
+from .core import load_plugins as _
 from .core.defaults import f
 from .core.import_names_conflict import (
     handle_import_names_conflict as _handle_import_names_conflict,
