+++ datar/core/plugin.py
@@ -2,9 +2,6 @@
 from typing import Any, List, Mapping, Tuple, Callable
 
 from simplug import Simplug, SimplugResult, makecall
-from pipda import register_array_ufunc
-
-from .options import get_option
 
 plugin = Simplug("datar")
 
@@ -17,19 +14,6 @@ def _collect(calls: List[Tuple[Callable, Tuple, Mapping]]) -> Mapping[str, Any]:
         if out is not None:
             collected.update(out)
     return collected
-
-
-def _array_ufunc_to_register(ufunc, x, *args, **kwargs):
-    """Register the array ufunc to pipda"""
-    from ..apis.other import array_ufunc
-
-    return array_ufunc(
-        x,
-        ufunc,
-        *args,
-        **kwargs,
-        __backend=array_ufunc.backend,
-    )
 
 
 @plugin.spec
@@ -85,9 +69,3 @@ def c_getitem(item):
 @plugin.spec(result=SimplugResult.SINGLE)
 def operate(op: str, x: Any, y: Any = None):
     """Operate on x and y"""
-
-
-plugin.load_entrypoints(only=get_option("backends"))
-
-plugin.hooks.setup()
-register_array_ufunc(_array_ufunc_to_register)
