+++ datar/dplyr.py
@@ -1,4 +1,5 @@
-from .core.plugin import plugin as _plugin
+
+from .core.load_plugins import plugin as _plugin
 from .apis.dplyr import *
 
 locals().update(_plugin.hooks.dplyr_api())
